feat: refactor TitleLogos. add conversation section
