+++ src/components/Governance/Delegates/index.tsx
@@ -1,7 +1,7 @@
 import { Avatar, Container, Grid, IconButton, SvgIcon, Typography } from '@mui/material'
 import React from 'react'
-import TwitterIcon from '@/public/images/twitter-icon.svg'
-import DiscordIcon from '@/public/images/discord-icon.svg'
+import TwitterIcon from '@/public/images/twitter-icon-28x28.svg'
+import DiscordIcon from '@/public/images/discord-icon-28x28.svg'
 import css from './styles.module.css'
 import type { BaseBlock } from '@/components/Home/types'
 import { shortenAddress } from '@/lib/shortenAddress'
