+++ src/components/common/BannerGradientImage/styles.module.css
@@ -1,7 +1,3 @@
-.hideMobile {
-  display: none;
-}
-
 .container {
   text-align: center;
   border-radius: 24px;
@@ -40,10 +36,6 @@
 }
 
 @media (min-width: 600px) {
-  .hideMobile {
-    display: block;
-  }
-
   .container {
     padding: 104px;
   }
