+++ src/content/home.json
@@ -208,7 +208,7 @@
         }
       }
     ],
-    "component": "Home/TitleLogos"
+    "component": "common/TitleLogos"
   },
   {
     "items": [
